Select and return valid dummy outlier masks

Valid dummy outlier masks are integers whose bitwise AND with the given invalid_mask is zero. This function returns the subset of integers of the given vector that does not share any bits with the given invalid_mask.

select_valid_masks(all_outlier_masks, invalid_mask)

Arguments

all_outlier_masks
A vector of possible outlier masks (integers).

invalid_mask
An integer encoding the invalid columns.

Value

The valid outlier_masks as a vector of integers.

Examples

all_outlier_masks <- c(0, 1, 2, 3, 4, 5, 6, 7)
invalid_mask <- 1
autovarCore:::select_valid_masks(all_outlier_masks, invalid_mask)

#> [1] 0 2 4 6