Construct linear and quadratic trend columns

This function returns a matrix of linear and quadratic trends.

trend_columns(number_of_rows)

Arguments

number_of_rows
the number of rows in the input data set.

Value

A matrix with number_of_rows rows and 2 columns, one for linear trends and one for quadratic trends.

Examples

autovarCore:::trend_columns(10)

#>       index index2
#>  [1,]     1      1
#>  [2,]     2      4
#>  [3,]     3      9
#>  [4,]     4     16
#>  [5,]     5     25
#>  [6,]     6     36
#>  [7,]     7     49
#>  [8,]     8     64
#>  [9,]     9     81
#> [10,]    10    100